Move resolver to test package
