+++ test/helpers/components/operator/installation.go
@@ -9,7 +9,7 @@ import (
 	"os/exec"
 	"testing"
 
-	"github.com/Dynatrace/dynatrace-operator/src/kubeobjects"
+	"github.com/Dynatrace/dynatrace-operator/test/helpers/platform"
 	"github.com/Dynatrace/dynatrace-operator/test/project"
 	"sigs.k8s.io/e2e-framework/pkg/envconf"
 	"sigs.k8s.io/e2e-framework/pkg/features"
@@ -71,7 +71,7 @@ func installViaHelm(t *testing.T, releaseTag string, withCsi bool, namespace str
 		helm.WithVersion(releaseTag),
 		helm.WithArgs("--create-namespace"),
 		helm.WithArgs("--install"),
-		helm.WithArgs("--set", fmt.Sprintf("platform=%s", kubeobjects.NewPlatformResolver().GetPlatform(t))),
+		helm.WithArgs("--set", fmt.Sprintf("platform=%s", platform.NewResolver().GetPlatform(t))),
 		helm.WithArgs("--set", "installCRD=true"),
 		helm.WithArgs("--set", fmt.Sprintf("csidriver.enabled=%t", withCsi)),
 		helm.WithArgs("--set", "manifests=true"),
