+++ test/helpers/proxy/proxy.go
@@ -8,10 +8,10 @@ import (
 	"testing"
 
 	dynatracev1beta1 "github.com/Dynatrace/dynatrace-operator/src/api/v1beta1"
-	"github.com/Dynatrace/dynatrace-operator/src/kubeobjects"
 	"github.com/Dynatrace/dynatrace-operator/test/helpers/kubeobjects/deployment"
 	"github.com/Dynatrace/dynatrace-operator/test/helpers/kubeobjects/manifests"
 	"github.com/Dynatrace/dynatrace-operator/test/helpers/kubeobjects/namespace"
+	"github.com/Dynatrace/dynatrace-operator/test/helpers/platform"
 	"github.com/Dynatrace/dynatrace-operator/test/helpers/sampleapps"
 	"github.com/Dynatrace/dynatrace-operator/test/project"
 	"sigs.k8s.io/e2e-framework/pkg/envconf"
@@ -49,7 +49,7 @@ func SetupProxyWithTeardown(t *testing.T, builder *features.FeatureBuilder, test
 }
 
 func installProxySCC(builder *features.FeatureBuilder, t *testing.T) {
-	if kubeobjects.NewPlatformResolver().IsOpenshift(t) {
+	if platform.NewResolver().IsOpenshift(t) {
 		builder.Assess("install proxy scc", manifests.InstallFromFile(proxySCCPath))
 	}
 }
